Implement the general layout of the web page
The layout of the sign up page was chosen to be split into two main
sections. The sign up page is supposed to have a splash logo section
on the left-hand side of the page and the sign up form on the right.

The splash logo section was cleaned up and finalized using placeholder
text and images.