Add more text input fields to the form
The input fields added were last name, email, and phone number.
First name, last name, and email are set to be required fields since I
felt they were going to be important information to keep track of for
any particular user.

Phone number is not required, but if entered, there is a pattern
validation for the user input in order to make sure the user doesn't
enter an invalid United States phone number.